requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dj-database-url==0.5.0
Django==3.1.6
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-extensions==3.1.0
django-js-asset==1.2.2
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
python-decouple==3.4
pytz==2021.1
regex==2020.11.13
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
whitenoise==5.2.0
youtube-dl==2021.2.22
